Update I2C_READ_CONTINUOUSLY to match I2C_READ w/ no provided slaveRegister

+++ utility/I2CFirmata.h
@@ -36,7 +36,7 @@
 /* i2c data */
 struct i2c_device_info {
   byte addr;
-  byte reg;
+  int reg;
   byte bytes;
 };
